adding more components

+++ Lionheart/lib/1240/components/overworld/OverworldMapComponent.h
@@ -6,7 +6,7 @@
 
 struct OverworldMapComponent : public ComponentBase
 {
-	double* heightMap_;
+	float* heightMap_;
 	int length_;
 	int width_;
 };
